Extract reporting to a module
