+++ src/test.js
@@ -1,4 +1,5 @@
 import assert from "assert";
+import report from "./reporters/default.js";
 
 export default addTest;
 
@@ -34,24 +35,4 @@ function runTest(test) {
     is: assert.strictEqual
   };
   return test.implementation(t);
-}
-
-async function report(runningTests, startTime) {
-  console.log("\nRunning...");
-  await printSequentially(runningTests);
-  console.log(`Ran in: ${Date.now() - startTime}ms`);
-}
-
-async function printSequentially(runningTests) {
-  for await (const result of runningTests) {
-    print(result);
-  }
-}
-
-function print(result) {
-  const { error, title } = result;
-  console.log(`${error ? "✗" : "✓"}  ${title}`);
-  if (error) {
-    console.log(error.stack);
-  }
 }
